+++ src/net/mod.rs
@@ -1,3 +1,3 @@
-pub mod parser;
+pub mod request;
 pub mod tcp;
 pub mod util;
